epub: fix inline image alpha compositing for transparent PNGs
InlineImageCacheEntry.pixels changed from RGB565 (pre-composited against
a fixed background) to ARGB8888. Per-pixel alpha blending now happens at
blit time against the actual screen pixels, so RGBA PNGs composite
correctly instead of rendering with a white background.

Cache key no longer includes bg565 since compositing is deferred.

+++ include/book/book.h
@@ -83,7 +83,7 @@ public:
     u8 layout_mode;
     u16 width;
     u16 height;
-    std::vector<u16> pixels;
+    std::vector<u32> pixels;  // ARGB8888: (a8<<24)|(r8<<16)|(g8<<8)|b8
   };
 
 private:
